app/requirements.txt
uvicorn
fastapi
python-multipart
transformers
Pillow
torch